Add `Public types` section to reactor.api.txt generator (#416)
Closes #416.

The canonical agent API index (`skills/reactor.api.txt` + its mirror under
`plugins/reactor/skills/reactor-dsl/references/`) previously only catalogued
5 sections — Factories, Modifiers, Hooks, Theme tokens, Enums — so newly-added
public surface like `WindowSpec.Opacity` / `ReactorWindow.SetPosition` was
invisible from the index alone (the issue's motivating example).

Adds a 6th `Public types` section that emits, per public non-static type,
public constructors / properties / methods / events (alphabetized, one line
per member, diffable). The 5 original sections are emitted byte-for-byte
unchanged (locked by `ExistingSections_Unchanged`).

ARM64-safe design: the existing generator was an apphost that crashes on
ARM64 (0xC0000005, the reason `-p:SkipReactorApiGen=true` exists). The
emit logic is extracted into a new sibling classlib `tools/Reactor.ApiIndex`
with no AfterBuild target; `Reactor.SignaturesGen` is now a thin shell. An
xUnit test (`Index_IsUpToDate`) drives `ApiIndexGenerator.Generate` in-process
on ARM64, acts as the PR-time staleness guard, and `UPDATE_API_INDEX=1`
regenerates both committed copies — the ARM64-safe regen path.

Tests: 6 new in `Tooling/ApiIndexGeneratorTests` (all green); full suite
9286/0/64 on ARM64.
