+++ tools/Reactor.ApiIndex/Reactor.ApiIndex.csproj
@@ -1,0 +1,24 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0-windows10.0.22621.0</TargetFramework>
+    <Platforms>x64;ARM64</Platforms>
+    <RootNamespace>Microsoft.UI.Reactor.ApiIndex</RootNamespace>
+    <AssemblyName>Reactor.ApiIndex</AssemblyName>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <UseWinUI>true</UseWinUI>
+    <WindowsAppSDKSelfContained>false</WindowsAppSDKSelfContained>
+    <WindowsPackageType>None</WindowsPackageType>
+    <IsPackable>false</IsPackable>
+    <NoWarn>$(NoWarn);CS1591</NoWarn>
+  </PropertyGroup>
+
+  <!-- Reflects over the built Reactor surface to build reactor.api.txt. This is a
+       plain class library with NO AfterBuild target, so referencing it never fires
+       the crashing apphost in tools/Reactor.SignaturesGen on ARM64. -->
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Reactor\Reactor.csproj" />
+  </ItemGroup>
+
+</Project>
