+++ tools/Reactor.SignaturesGen/Reactor.SignaturesGen.csproj
@@ -13,9 +13,11 @@
     <NoWarn>$(NoWarn);CS1591</NoWarn>
   </PropertyGroup>
 
-  <!-- Reflect over the freshly-built Reactor surface to emit reactor.api.txt. -->
+  <!-- Reflect over the freshly-built Reactor surface to emit reactor.api.txt. The
+       index text is built by the sibling Reactor.ApiIndex classlib. -->
   <ItemGroup>
     <ProjectReference Include="..\..\src\Reactor\Reactor.csproj" />
+    <ProjectReference Include="..\Reactor.ApiIndex\Reactor.ApiIndex.csproj" />
   </ItemGroup>
 
   <!-- After build, run the generator with the repo root as argv[0] so the file
